+++ app/src/main/res/layout/fragment_event_1.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="endorphine.icampyou.EventFragment1"
+    tools:context="endorphine.icampyou.EventMenu.EventFragment1"
     android:background="#FFFFFF">
 
     <TextView
