+++ app/src/main/res/layout/fragment_home_1.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="endorphine.icampyou.HomeFragment1"
+    tools:context="endorphine.icampyou.HomeMenu.HomeFragment1"
     android:background="#FFFFFF">
 
     <TextView
